+++ oda_reader/crs.py
@@ -1,7 +1,14 @@
+from pathlib import Path
+
 import pandas as pd
 
 from oda_reader.common import logger
-from oda_reader.download.download_tools import download
+from oda_reader.download.download_tools import (
+    get_bulk_file_id,
+    bulk_download_parquet,
+    CRS_FLOW_URL,
+    download,
+)
 
 DATAFLOW_ID: str = "DSD_CRS@DF_CRS"
 
@@ -9,8 +16,70 @@ DATAFLOW_ID: str = "DSD_CRS@DF_CRS"
 {donor}.{recipient}.{sector}.{measure}.{channel}.
         {modality}.{flow_type}.{price_base}.{md_dim}.{md_id}.{unit_measure}.
         {time_period}
-
 """
+
+def get_full_crs_parquet_id():
+    return get_bulk_file_id(flow_url=CRS_FLOW_URL, search_string="CRS-Parquet|")
+
+
+def get_reduced_crs_parquet_id():
+    return get_bulk_file_id(flow_url=CRS_FLOW_URL, search_string="CRS-reduced-parquet|")
+
+
+def get_year_crs_zip_id(year: int):
+    return get_bulk_file_id(
+        flow_url=CRS_FLOW_URL, search_string=f"CRS {year} (dotStat format)|"
+    )
+
+
+def download_crs_file(year: int, save_to_path: Path | str | None = None):
+    """
+    Download a year of CRS data from the bulk download service. The file is large.
+    It is therefore strongly recommended to save it to disk. If save_to_path is not
+    provided, the function will return a DataFrame.
+
+    Args:
+        year: The year of CRS data to download.
+        save_to_path: The path to save the file to. Optional. If not provided, a
+        DataFrame is returned.
+
+    Returns:
+        pd.DataFrame | None: The DataFrame if save_to_path is not provided.
+
+    """
+
+    file_id = get_year_crs_zip_id(year=year)
+
+    return bulk_download_parquet(
+        file_id=file_id, save_to_path=save_to_path, is_txt=True
+    )
+
+
+def bulk_download_crs(
+    save_to_path: Path | str | None = None, reduced_version: bool = False
+):
+    """
+    Bulk download the CRS data from the bulk download service. The file is very large.
+    It is therefore strongly recommended to save it to disk. If save_to_path is not
+    provided, the function will return a DataFrame.
+
+    Args:
+        save_to_path: The path to save the file to. Optional. If not provided, a
+        DataFrame is returned.
+        reduced_version: Whether to download the reduced version of the CRS data.
+
+    Returns:
+        pd.DataFrame | None: The DataFrame if save_to_path is not provided.
+
+    """
+
+    if reduced_version:
+        file_id = get_reduced_crs_parquet_id()
+    else:
+        file_id = get_full_crs_parquet_id()
+
+    return bulk_download_parquet(file_id=file_id, save_to_path=save_to_path)
+
 
 def download_crs(
     start_year: int | None = None,